e refactoring
