+++ go.mod
@@ -3,8 +3,3 @@ module github.com/inpure/go2docker
 go 1.17
 
 require github.com/spf13/cobra v1.4.0
-
-require (
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-)
