+++ pkg/hello/sayHello.go
@@ -1,0 +1,13 @@
+package hello
+
+import (
+	"log"
+	"time"
+)
+
+func LoopSay(name string, pause int) {
+	for {
+		log.Println("Hello " + name)
+		time.Sleep(time.Duration(pause) * time.Second)
+	}
+}
